updated card block in index.handlebars, removed unordered list
