+++ package.json
@@ -7,7 +7,7 @@
     "start": "node server.js",
     "lint": "eslint **/*.js --quiet",
     "fix": "eslint --fix .",
-    "test": "npm run lint && NODE_ENV=test mocha -u tdd --reporter spec --exit"
+    "test": "npm run lint && SET NODE_ENV=test && mocha -u tdd --reporter spec --exit"
   },
   "license": "ISC",
   "dependencies": {
