+++ routes/htmlRoutes.js
@@ -1,8 +1,25 @@
+var db = require("../models");
+
 module.exports = function(app) {
-  // Load index page
+  // Get most recent (default setting)
   app.get("/", function(req, res) {
-    res.render("index");
+    var query = {};
+    if (req.query.user_id) {
+      query.UserId = req.query.user_id;
+    }
+    db.Entries.findAll({
+      include: [db.User],
+      where: query
+    }).then(function(data) {
+      var hbsObject = {
+        entries: data
+      };
+      res.render("index", hbsObject);
+      // console.log(hbsObject);
+    });
   });
+
+  // Get most popular
 
   // Render 404 page for any unmatched routes
   app.get("*", function(req, res) {
