+++ views/index.handlebars
@@ -48,11 +48,11 @@
 <div class="chalkboard">
   <!-- <h1>{{msg}}</h1> -->
   <h2>Welcome To StouijiBoard!</h2>
-<<<<<<< HEAD
-      
-<hr>
+
+  <hr>
+
+
   
-  <div>
     <div id="renderCard" class="card">
       {{#each entries}}
       <div class="card-header">
@@ -65,46 +65,8 @@
             <a href="#" class="btn btn-primary">Contribute to Story</a>
       </div>
       {{/each}}
+      <br>
     </div>
-    <div class="card">
-      <div class="card-header">
-      Created by: xxxxx, created on: xxxxxx, Number of posts: xxxxx, last contributed: xxxxxxx
-      </div>
-      <div class="card-body">
-        <h4 class="card-title">Now Is The Time For All Good Men To Come To The Aid Of Their Country</h4>
-        <p class="card-text">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed condimentum lectus a nibh
-          eleifend, eu pharetra
-          lectus tempor. In auctor fringilla dolor eleifend lobortis. Morbi mollis dui mauris, at accumsan lorem
-          ultricies vel. Sed sagittis ultricies ultrices. Mauris aliquam et arcu vel efficitur. Nunc eget ultrices
-          purus,
-          eget venenatis diam. Duis aliquet cursus vulputate. Ut elementum euismod lectus, eget euismod tortor
-          feugiat
-          eu. Sed ut porttitor dolor. Pellentesque pretium libero sit amet ligula semper, nec venenatis dui commodo.
-          Proin commodo, nisl eget pulvinar gravida, magna nisl blandit nibh, quis dapibus dui ipsum in erat. Etiam
-          fermentum quam ac ex venenatis, ac tincidunt tellus commodo. Duis facilisis nisl sit amet egestas pharetra.
-          Donec vestibulum ante et massa varius, ut auctor diam scelerisque.</p>
-        <a href="#" class="btn btn-primary">Go somewhere</a>
-      </div>
-    </div>
-    <div>
-      <p>
-        Ut iaculis sollicitudin tincidunt. Donec posuere orci id sollicitudin imperdiet. Pellentesque habitant morbi
-        tristique senectus et netus et malesuada fames ac turpis egestas. Duis nunc felis, condimentum nec elit
-        malesuada, rhoncus scelerisque felis. Integer imperdiet orci vitae metus tempus, ut pretium nunc gravida.
-        Donec
-        tempus vulputate massa, vel mollis purus tristique in. Phasellus varius feugiat nunc. Aenean non quam ac est
-        condimentum efficitur. Phasellus dapibus eget urna sollicitudin tincidunt.
-=======
->>>>>>> master
-
-  <hr>
-
-
-  <div class="card">
-    <ul>
-      {{#each entries}} {{> entries/story-block}} {{/each}}
-    </ul>
-  </div>
   <div class="card">
     <ul>
       <div class="card-header">
