+++ views/partials/entries/story-block.handlebars
@@ -1,9 +1,9 @@
 <div class="card-header">
-      Created by: {{author}}, created on: {{createdAt}}, last contributed: {{updatedAt}}
+      Created by: {{author}}, Created on: {{createdAt}}, Number of posts: xxxxx, Last contributed: {{updatedAt}}
 </div>
 <div id="renderCard" class="card-body">
       <h4 class="card-title">{{title}}</h4>
       <p class="card-text">{{entry}}</p>
-      <a href="#" class="btn btn-primary"></a>
+      <a href="#" class="btn btn-primary">Contribute to the Story</a>
 </div>
